third_party/blink/web_tests/fast/events/mouseenter-mouseleave-chained-listeners-expected.txt

// Entering #inner1
mouseover (at target) on inner1
mouseover (bubbling) on inner1
mouseenter (at target) on outer
mouseenter (at target) on inner1
// Entering #outer and #inner2
mouseout (at target) on inner1
mouseout (bubbling) on inner1
mouseleave (at target) on inner1
// Setting capturing mouseenter handler on #outer from #inner1's mouseleave handler.
mouseover (at target) on inner2
mouseover (bubbling) on inner2
mouseenter (capturing) on inner2
PASS successfullyParsed is true

TEST COMPLETE
